Fix HateSpeech and Offensive dataset names
