+++ assets/ar/factuality_disinformation_harmful_content/hate_speech/OSACT4SubtaskB_GPT35_ZeroShot.py
@@ -1,0 +1,47 @@
+import os
+
+from llmebench.datasets import OSACT4SubtaskBDataset
+from llmebench.models import GPTModel, RandomGPTModel
+from llmebench.tasks import HateSpeechTask
+
+
+def config():
+    return {
+        "dataset": OSACT4SubtaskBDataset,
+        "dataset_args": {},
+        "task": HateSpeechTask,
+        "task_args": {},
+        "model": GPTModel,
+        "model_args": {
+            "api_type": "azure",
+            "api_version": "2023-03-15-preview",
+            "api_base": os.environ["AZURE_API_URL"],
+            "api_key": os.environ["AZURE_API_KEY"],
+            "engine_name": os.environ["ENGINE_NAME"],
+            "class_labels": ["HS", "NOT_HS"],
+            "max_tries": 3,
+        },
+        "general_args": {
+            "data_path": "data/factuality_disinformation_harmful_content/hate_speech/OSACT2020-sharedTask-test-tweets-labels.txt"
+        },
+    }
+
+
+def prompt(input_sample):
+    return {
+        "system_message": "You are an AI assistant that helps people find information.",
+        "messages": [
+            {
+                "sender": "user",
+                "text": f'if the following Arabic sentence has hate speech, just say "HS", otherwise, say just "NOT_HS" without explanation: \n {input_sample}',
+            }
+        ],
+    }
+
+
+def post_process(response):
+    out = response["choices"][0]["text"]
+    j = out.find(".")
+    if j > 0:
+        out = out[0:j]
+    return out
