+++ assets/ar/factuality_disinformation_harmful_content/hate_speech/OSACT4SubtaskB_GPT4_FewShot.py
@@ -1,15 +1,15 @@
 import os
 
-from llmebench.datasets import OffensiveDataset
+from llmebench.datasets import OSACT4SubtaskBDataset
 from llmebench.models import GPTChatCompletionModel
-from llmebench.tasks import OffensiveTask
+from llmebench.tasks import HateSpeechTask
 
 
 def config():
     return {
-        "dataset": OffensiveDataset,
+        "dataset": OSACT4SubtaskBDataset,
         "dataset_args": {},
-        "task": OffensiveTask,
+        "task": HateSpeechTask,
         "task_args": {},
         "model": GPTChatCompletionModel,
         "model_args": {
@@ -18,13 +18,12 @@ def config():
             "api_base": os.environ["AZURE_API_URL"],
             "api_key": os.environ["AZURE_API_KEY"],
             "engine_name": os.environ["ENGINE_NAME"],
-            "class_labels": ["OFF", "NOT_OFF"],
             "max_tries": 3,
         },
         "general_args": {
-            "data_path": "data/factuality_disinformation_harmful_content/offensive_language/OSACT2020-sharedTask-test-tweets-labels.txt",
+            "data_path": "data/factuality_disinformation_harmful_content/hate_speech/OSACT2020-sharedTask-test-tweets-labels.txt",
             "fewshot": {
-                "train_data_path": "data/factuality_disinformation_harmful_content/offensive_language/OSACT2020-sharedTask-train_OFF.txt",  # TO_DO
+                "train_data_path": "data/factuality_disinformation_harmful_content/hate_speech/OSACT2020-sharedTask-train_HS.txt",  # TO_DO
             },
         },
     }
@@ -32,14 +31,14 @@ def config():
 
 def prompt(input_sample, examples):
     base_prompt = (
-        "Given a tweet, predict whether it is offensive or not. Answer only by using"
-        " offensive and not_offensive. Here are some examples:\n"
+        "Given a tweet, predict whether it contains hate speech. Answer only by using"
+        " hate_speech and not_hate_speech. Here are some examples:\n"
     )
 
     return [
         {
             "role": "system",
-            "content": "You are an expert annotator, you can identify and label offensive content within a tweet.",
+            "content": "You are an expert annotator, you can identify and label hate speech content within a tweet.",
         },
         {
             "role": "user",
@@ -53,14 +52,13 @@ def few_shot_prompt(input_sample, base_prompt, examples):
     for example in examples:
         # Found chatgpt confused when using 0 and 1 in the prompt
         # label = "no" if example["label"] == "0" else "yes"
-        label = "not_offensive" if example["label"] == "NOT_OFF" else "offensive"
-
+        label = "not_hate_speech" if example["label"] == "NOT_HS" else "hate_speech"
         out_prompt = (
             out_prompt + "Tweet: " + example["input"] + "\nLabel: " + label + "\n\n"
         )
 
     # Append the sentence we want the model to predict for but leave the Label blank
-    out_prompt = out_prompt + "Tweet: " + input_sample + "\nLabel: \n"
+    out_prompt = out_prompt + "Tweet: " + input_sample + "\nLabel:\n"
 
     # print("=========== FS Prompt =============\n")
     # print(out_prompt)
@@ -74,10 +72,10 @@ def post_process(response):
     if j > 0:
         out = out[0:j]
 
-    if "not_offensive" in out:
-        out = "NOT_OFF"
-    elif "offensive" in out:
-        out = "OFF"
+    if "not_hate_speech" in out or "no_hate_speech" in out:
+        out = "NOT_HS"
+    elif "hate_speech" in out:
+        out = "HS"
     else:
         out = None
     return out
