--- assets/ar/factuality_disinformation_harmful_content/offensive_language/Offensive_GPT4_ZeroShot.py
@@ -1,48 +1,0 @@
-import os
-
-from llmebench.datasets import OffensiveDataset
-from llmebench.models import GPTChatCompletionModel
-from llmebench.tasks import OffensiveTask
-
-
-def config():
-    return {
-        "dataset": OffensiveDataset,
-        "dataset_args": {},
-        "task": OffensiveTask,
-        "task_args": {},
-        "model": GPTChatCompletionModel,
-        "model_args": {
-            "api_type": "azure",
-            "api_version": "2023-03-15-preview",
-            "api_base": os.environ["AZURE_API_URL"],
-            "api_key": os.environ["AZURE_API_KEY"],
-            "engine_name": os.environ["ENGINE_NAME"],
-            "class_labels": ["OFF", "NOT_OFF"],
-            "max_tries": 3,
-        },
-        "general_args": {
-            "data_path": "data/sentiment_emotion_others/offensive_language/OSACT2020-sharedTask-test-tweets-labels.txt"
-        },
-    }
-
-
-def prompt(input_sample):
-    return [
-        {
-            "role": "system",
-            "content": "You are an AI assistant that helps people find information on locations.",
-        },
-        {
-            "role": "user",
-            "content": f'if the following Arabic sentence is offensive, just say "OFF", otherwise, say just "NOT_OFF" without explanation: \n {input_sample}',
-        },
-    ]
-
-
-def post_process(response):
-    out = response["choices"][0]["message"]["content"]
-    j = out.find(".")
-    if j > 0:
-        out = out[0:j]
-    return out
